manifold-sql changes - add support for direct access to one-many and many-many fks from table rows e.g., film.getCategoryRefs() / category.getFilmRefs() etc. - crud spi supports readOne/readMany in terms of fks - re-added fk id columns to table rows e.g., both city.getCountryId() & city.getCountryRef(), for perf and direct ids are generally more useful than I had thought (as opposed to fetching the ref and getting the id) - improved fk naming scheme to be more concise and remain distinct - build/create method overload parameter lists are ordered consistently wrt refs vs. ids - replaced anon table row creation with named Impl class as inner class to table row interfaces, easier on the eye in the debugger and generally better - replaced usage of ConcurrentHashMap/Set to LinkedHashMap/Set where concurrency was not needed and where natural order is useful - fixed giant perf issue with SchemaManifold#getTypeNameForFile, particularly for IJ

+++ manifold-deps-parent/manifold-sql-test/src/test/java/manifold/sql/schema/crud/CrudTest.java
@@ -43,7 +43,7 @@ public class CrudTest extends H2SakilaTest
 
     auto row = "[.sql:H2Sakila/] SELECT country_id FROM Country where country = 'mycountry'"
       .fetchOne( txScope );
-    assertEquals( row.getCountryId(), hi.getCountryId() );
+    assertEquals( row.getCountryId(), (long)hi.getCountryId() );
   }
 
   @Test
